+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.3.2 (Next)
 - Miscellaneous minor Terraform module plugin fixes.
 - Minor documentation corrections.
+- Update GoSS `format` and Packer `on_error` validations.
 
 ### 1.3.1
 - Improve accuracy of `default` in plugin documentation.